[2023.02.16 노영준] 뭐였지

+++ app/src/main/java/com/choonham/lck_manager/map_object/Dragon.java
@@ -1,7 +1,7 @@
 package com.choonham.lck_manager.map_object;
 
 public class Dragon extends MapObject{
-    public Dragon(String objectName, float x, float y, int visionDistance, int teamSide) {
+    public Dragon(String objectName, int x, int y, int visionDistance, int teamSide) {
         super(objectName, x, y, visionDistance, teamSide);
     }
 }
