+++ app/src/main/java/com/choonham/lck_manager/map_object/MapObject.java
@@ -4,15 +4,15 @@ public class MapObject {
 
     public String objectName;
 
-    public float x;
-    public float y;
+    public int x;
+    public int y;
 
     public int visionDistance;
 
     // 0: red, 1: blue
     public int teamSide;
 
-    public MapObject(String objectName, float x, float y, int visionDistance, int teamSide) {
+    public MapObject(String objectName, int x, int y, int visionDistance, int teamSide) {
         this.objectName = objectName;
         this.visionDistance = visionDistance;
         this.teamSide = teamSide;
