+++ app/src/main/java/com/choonham/lck_manager/map_object/Nexus.java
@@ -3,7 +3,7 @@ package com.choonham.lck_manager.map_object;
 public class Nexus extends MapObject{
     private int playerSide;
 
-    public Nexus(String objectName, float x, float y, int visionDistance, int teamSide, int playerSide) {
+    public Nexus(String objectName, int x, int y, int visionDistance, int teamSide, int playerSide) {
         super(objectName, x, y, visionDistance, teamSide);
 
         this.playerSide = playerSide;
