+++ app/src/main/java/com/choonham/lck_manager/map_object/Turret.java
@@ -4,7 +4,7 @@ public class Turret  extends MapObject{
 
     private int playerSide;
 
-    public Turret(String objectName, float x, float y, int visionDistance, int teamSide, int playerSide) {
+    public Turret(String objectName, int x, int y, int visionDistance, int teamSide, int playerSide) {
         super(objectName, x, y, visionDistance, teamSide);
 
         this.playerSide = playerSide;
